requirements.txt
asgiref==3.8.1
boto3==1.34.121
botocore==1.34.121
dj-database-url==2.2.0
Django==5.0.6
django-storages==1.14.3
gunicorn==22.0.0
jmespath==1.0.1
packaging==24.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.1
urllib3==2.2.1
